[Xamarin.Android.Build.Tasks] Fix aapt_rules.txt corruption (#7587)
Fixes: https://github.com/xamarin/xamarin-android/issues/7447

Users are reporting a error with the aapt_rules file generated by aapt2

	MSBUILD : java.exe error JAVA0000: Error in obj\Release\110\aapt_rules.txt at line 289, column 1: [C:\Jenkins\workspace\Demo\Demo.Android\Demo.Android.csproj]
	MSBUILD : java.exe error JAVA0000: Expected char '-' at obj\Release\110\aapt_rules.txt:289:1 [C:\Jenkins\workspace\Demo\Demo.Android\Demo.Android.csproj]
	MSBUILD : java.exe error JAVA0000: nit>(...); } [C:\Jenkins\workspace\Demo\Demo.Android\Demo.Android.csproj]
	MSBUILD : java.exe error JAVA0000: Compilation failed

This only happens when building with
`$(AndroidCreatePackagePerAbi)`=True.   After looking at the code:

 1. we call `aapt2 link` multiple times when building per abi
 2. the temp file used for the `aapt2 link` proguard output was the
    same file for *all* of the `aapt2 link` calls.

The combination of (1) and (2) can result in one call to `aapt2`
overwriting the previous `aapt2` invocation results, or (worse)
leaving garbage at the end of the file.
results... or worse leaving garbage at the end.

Fix this by providing a different temp file for each call to `aapt2`.
These files are then combined later just in case they contain unique
information for that abi.  While this does result in duplicate lines
in the `aapt_rules` file, this does not impact on the build.