[Xamarin.Android.Build.Tasks] fix cases of missing `@(Reference)`
Fixes: https://github.com/dotnet/maui/issues/10154

If you have a solution setup with:

* `ApplicationA` project reference ->
* `LibraryB` reference ->
* `LibraryC`

The app will crash at runtime, due to a missing `LibraryC.dll`.

You can solve the problem by changing `@(Reference)` to a
`@(ProjectReference)`. However, it appears the same situation works in
a .NET 7 self-contained console app:

    dotnet publish --self-contained -r win-x64
    ...
    ls -n .\bin\Debug\net7.0\win-x64\publish\LibraryC.dll
    LibraryC.dll

The underlying issue appears to be:

https://github.com/dotnet/msbuild/blob/a2490dd3f78cce4abc8f9e6f1b5268437332818f/src/Tasks/Microsoft.Common.CurrentVersion.targets#L2322

In the console app, `$(BuildingProject)` is `true` and
`$(_FindDependencies)` is empty.

In the android app, `$(BuildingProject)` is `false` and
`$(_FindDependencies)` is `false`.

It appears the `BuildOnlySettings` target *should* be running in
Android apps when we do an "inner" build per `$(RuntimeIdentifier)`.
Simply adding a dependency for the `_ComputeFilesToPublishForRuntimeIdentifiers`
MSBuild target solves this issue.

This likely might fix other issues, such as:

https://github.com/dotnet/maui/issues/11364

I added a test for this scenario.