[Mono.Android] ServerCertificateValidationCallback() and redirects (#7662)
Fixes: https://github.com/xamarin/xamarin-android/issues/7650

`AndroidMessageHandler` was incorrectly validating SSL certificate
after a redirect to another domain if and only if
`AndroidMessageHandler.ServerCertificateCustomValidationCallback`
was set; consider:

	const string url = "https://httpbin.org/redirect-to?url=https://www.microsoft.com/";
	using var handler = new AndroidMessageHandler() {
	    ServerCertificateCustomValidationCallback = (message, certificate, chain, errors) => errors == SslPolicyErrors.None
	};
	using var client = new HttpClient(handler);
	await client.GetAsync(url);

The expectation is that
`AndroidMessageHandler.ServerCertificateCustomValidationCallback`
will be called *twice*:

 1. Once with `message.RequestUri` set to
    `https://httpbin.org/redirect-to?…`, and
 2. Once again with `message.RequestUri` set to
    `https://www.microsoft.com`.

The problem was that in actuality the 2nd invocation was unchanged
from (1), with `message.RequestUri` set to
`https://httpbin.org/redirect-to…?…`, and `errors` set to
`SslPolicyErrors.RemoteCertificateNameMismatch` (!).

The problem was that `AndroidMessageHandler.DoSendAsync()` didn't
update `HttpRequestMessage.RequestUri` to the redirected URI, causing
the `errors` enum value to contain an error, and the second
`ServerCertificateCustomValidationCallback` invocation to have the
wrong information.

Update `AndroidMessageHandler.DoSendAsync()` so that when dealing
with an HTTP redirect, `message.RequestUri` always contains the
"current" value in the redirect chain.  This keeps `RequestUri` in
sync with the redirect status to avoid this problem.