[Xamarin.Android.Build.Tasks] skip XA1034 logic in some cases (#7680)
After profiling a build for a `dotnet new maui` app, I noticed:

    193.41ms xamarin.android.build.tasks!Xamarin.Android.Tasks.MonoAndroidHelper.HasResourceDesignerAssemblyReference()

Which equates to:

    Top 10 most expensive tasks
    ...
    AppendCustomMetadataToItemGroup = 204 ms

In dc3ccf2, we added this check to emit `XA1034` when
`$(AndroidUseDesignerAssembly)` is false. And so we don't actually
need to validate this value when it is `true` (the default).

I moved the logic from the `<AppendCustomMetadataToItemGroup/>`
MSBuild task to a new `<CheckForInvalidDesignerConfig/>` task. In most
projects, it will be completely skipped:

    Target "_CheckForInvalidDesignerConfig" skipped, due to false condition; ( '$(AndroidUseDesignerAssembly)' != 'True' ) was evaluated as ( 'true' != 'True' ).

After these changes, I get a much better number in projects with
default settings:

    AppendCustomMetadataToItemGroup = 2 ms

Just by moving the task around, it is faster in general:

    CheckForInvalidDesignerConfig = 11 ms

This is due to less assemblies passed in than before:

    Assemblies="@(_MonoAndroidReferencePath);@(_MonoAndroidReferenceDependencyPaths)"

While previously, we were checking *all* assemblies.