[WIP] [Xamarin.Android.Build.Tasks] use hashes as `Inputs` for `_GenerateJavaStubs`
One of the slower parts of an incremental Android build is:

    _GenerateJavaStubs = 978 ms

Currently, this happens when any .NET assembly changes: C# change,
`.xaml` change, etc.

In theory, we could analyze .NET assemblies during the linker and
create a "hash" that represents types & members that would invalidate
`_GenerateJavaStubs`'s output. We can then save these hashes in a file
to be used as `Inputs` instead of the .NET assemblies themselves.

Example of the hash:

    HashAlgorithm hasher = new Microsoft.Android.Build.Tasks.Crc64 ();
    foreach (var type in scanner.GetJavaTypes (assembly)) {
        Hash (hasher, type.FullName);
        foreach (var method in type.Methods) {
            Hash (hasher, method.Name);
        }
    }

The result is:

    Target Name=_GenerateJavaStubs Project=bar.csproj
    Skipping target "_GenerateJavaStubs" because all output files are up-to-date with respect to the input files.

Improving the incremental build time about about ~1 second in a
`dotnet new maui` project. `_LinkAssembliesNoShrink` will be slightly
slower, but because it builds partially it will only be creating this
hash on a single file most of the time. Currently, this seems to be
about ~20ms final numbers TBD.

Things left to do:

* Make sure we are hashing the right things that `_GenerateJavaStubs`
  cares about.

* Fix `Release` builds, they likely aren't setup right.

* Add several tests around this scenario. We need to validate adding
  new `Java.Lang.Object` subclasses work as expected.