Revert "[Xamarin.Android.Build.Tasks] fix cases of missing `@(Reference)` (#7642)" (#7726)
This reverts commit 22f2001.

.NET MAUI is hitting a build failure such as:

    Unable to open file 'obj\Release\net8.0-android\android-x64\aot\x86_64\Microsoft.Maui.Controls.resources\temp.s': Permission denied

In 22f2001, we began passing satellite assemblies to the AOT compiler,
on accident? I am unsure how a test in xamarin-android didn't catch
this, but it may be something that only happens via a
`@(ProjectReference)` as occurred in .NET MAUI.

For now, let's revert the change and revisit later.

We should reopen https://github.com/dotnet/maui/issues/10154 after
this is merged.