Replace all `exit()` calls with `abort()` in native code
`exit()` causes the application to restart, which can be both annoying
and confusing when the app seemingly shows something (usually a white
screen) and yet it doesn't work.  The only way to learn what's going on
is to look at logcat output to see the error message and a restart.

`abort()`, however, fully terminates the application and also shows a
native stack trace, which makes it easy to find the cause of the
"crash" (the error message will be logged typically just before the
native stack trace is printed)