Tweak AndroidMessageHandler behavior for WCF support
Context: https://github.com/xamarin/xamarin-android/issues/7230
Context: https://github.com/dotnet/runtime/issues/80935

When a WCF application receives compressed content that is automatically
decompressed in `AndroidMessageHandler`, it will most of the time fail
to properly read the response, cutting reading of the decompressed
content short.  The reason for this is that when `AndroidMessageHandler`
creates a wrapper decompression stream, it does not update the
`Content-Length` to match the length of the decoded content, because it
doesn't have a way to know what the length is without first reading the
stream to the end, and that might prevent the end user to read the
content.  Additionally, I think the `Content-Length` header should
reflect the **original** content length, for the end user to be able to
interpret the response as it was sent.

WCF, on the other hand, looks at the `Content-Length` header and, if
found, it takes its value and reads only this many bytes from the
content stream and no more - it will almost always result in short reads
and failure to correctly interpret the response.

I implemented this workaround which makes `AndroidMessageHandler` behave
the same way as other handlers implemented in the BCL.  What they do in
this situation, is to remove the `Content-Length` header, making WCF
read stream to the end.  Additionally, the clients remove the compressed
content encoding identifier from the `Content-Encoding` header.

As a bonus, this commit also adds support for decompression of responses
compressed with the `Brotli` compression (using the `br` encoding ID in
the `Content-Encoding` header)