Attempt to get better logcat logging for failed tests
I noticed that more often than not, we don't see the actual error in
`logcat-failed.log` when the test fails on CI, but when it runs locally,
we can see the issue logged properly.

I'm not 100% what causes it on CI, but I suspect that we might be facing
a problem with the logcat buffer wrapping around and overwriting the
interesting data.  This might be because of the fact that `logcat -c`
runs only once per **class** containing tests and also because we log
too quickly after the app quits.

To remedy the former, run `logcat -c` before each test starts and to
attempt a fix for the latter, pause for 1 second before dumping logcat
buffer after the test fails.