[Xamarin.Android.Build.Tasks] Improve aapt2+file not found handling (#7644)
Context: https://github.com/xamarin/xamarin-android/issues/7639

If a project has a long path name, this can cause `aapt2` to fail:

	error APT2000: The system cannot find the file specified. (2).
	
This is not very helpful since there is no obvious reason to the user
as to *why* the build is failing.

We should do better.  We *can* do better.

Catch this specific error message and report a helpful, actionable
message instead.

	error APT2264: This is probably caused by the project exceeding the Windows maximum path length limitation.
	  See https://learn.microsoft.com/xamarin/android/errors-and-warnings/apt2264 for details.


In this case we will advise the user that the path is probably too
long and they should consult the documentation on how to resolve it.

The documentation includes a number of options for resolutions, so it
should cover all use cases.  Generally these are:

 1. Move the project closer to the root of the drive.
 2. Enable long path support in windows.
 3. Make changes to the `$(BaseIntermediateOutputPath)` MSBuild
    property to move only the `obj` folder closer to the drive root.

Most users can get away with (1), though with Maui and Blazor based
apps this might not solve the problem as the project itself might
have a deep folder structure.  Options (2) and (3) are available to
cover those scenarios.