[monodroid] Properly process satellite assemblies
Fixes: https://github.com/xamarin/xamarin-android/issues/7819

Our native runtime uses a cache of pointers to loaded managed assembly
images (essentially an array of the native struct `MonoImage` pointers)
which is pre-allocated at build time and placed in the native
`libxamarin-app.so` library.

While generating the cache, we also generate hashes for a number of
assembly name permutations (currently two per assembly - with and
without the extension).  Only unique assembly names are considered when
generating the cache (it's possible to have duplicate names, because we
package more than one copy of some assemblies - those which are
architecture specific).  This algorithm had a bug which made it ignore
culture prefix in satellite assembly names (e.g.
`en/MyAssembly.resources.dll`) and instead of several entries for each
culture, we generated only two entries for `MyAssembly.resources.dll`
and `MyAssembly.resources` but we still counted each culture-prefixed
assembly and stored that number in `libxamarin-app.so` to be used at
runtime to calculate number of entries in the cache.

This made the array storing cached `MonoImage*` pointers to be smaller
than the number of actual assemblies in the APK times 2 and in some
cases we failed to look up pointer to some images and, as the result,
passed a `NULL` pointer to MonoVM which then caused a segmentation fault
trying to dereference the pointer.

Stop ignoring the culture prefix for satellite assemblies in order to
avoid the situation.  Additionally, since the previous assumption that
MonoVM will validate all pointers passed to its APIs turned out to be
unwarranted, we now check more carefully for `NULL` pointers when trying
to obtain a native function pointer from the MonoVM runtime.